Working with South enabled!
South changes INSTALLED_APPS setting on the fly, so we need to first look up
which shared_apps and tenant_apps we need right now and only handle those
during the sync.
If the tenant model uses migrations, querying all the tenants will fail on
`sync_schemas` command because the tables for that model is not created yet.
(It is only created after south's `migrate` command.) I handle this now by
checking for ProgrammingError exception (because that's what Django throws if
you want to query a non-existing table) and warn the user to apply the migrations.