Logically reorganized code, fixed some logic mistakes.
Tried to put codes to the logically most fitting place;
shared_app models, tenant_apps models, and shared_models are now three different properties.
Select models to sync in sync_schemas.
If no SHARED_APPS, INSTALLED_APPS will be used.
If no TENANT_APPS, INSTALLED_APPS will be used.
If no SHARED_MODELS, connection.shared_models will be empty.
If a model is in SHARED_MODELS and it's app in TENANT_APPS,
the model's table will not be created on the tenant's schema,
but the foreign key will be forced to the public schema.