Create convert_to_south_schemas.py
convert_to_south_schemas is the management command to invoke South's convert_to_south command. This is required in order to convert an existing application that has existing tables to South. Use the command this way: manage.py convert_to_south_schemas myapp (where myapp is an application). They would follow this with the command: manage.py migrate_schemas myapp 0001 --fake  to fake the migration. Do this for all applications that you need to convert to South. Refer to the South documentation for more details.